+++ clients/pkg/promtail/targets/file/filetarget.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/grafana/loki/clients/pkg/promtail/api"
 	"github.com/grafana/loki/clients/pkg/promtail/positions"
+	"github.com/grafana/loki/clients/pkg/promtail/scrapeconfig"
 	"github.com/grafana/loki/clients/pkg/promtail/targets/target"
 )
 
@@ -75,6 +76,8 @@ type FileTarget struct {
 
 	targetConfig *Config
 
+	decompressCfg *scrapeconfig.DecompressionConfig
+
 	encoding string
 }
 
@@ -92,6 +95,7 @@ func NewFileTarget(
 	fileEventWatcher chan fsnotify.Event,
 	targetEventHandler chan fileTargetEvent,
 	encoding string,
+	decompressCfg *scrapeconfig.DecompressionConfig,
 ) (*FileTarget, error) {
 	t := &FileTarget{
 		logger:             logger,
@@ -109,6 +113,7 @@ func NewFileTarget(
 		fileEventWatcher:   fileEventWatcher,
 		targetEventHandler: targetEventHandler,
 		encoding:           encoding,
+		decompressCfg:      decompressCfg,
 	}
 
 	go t.run()
@@ -318,9 +323,9 @@ func (t *FileTarget) startTailing(ps []string) {
 		}
 
 		var reader Reader
-		if isCompressed(p) {
+		if t.decompressCfg != nil && t.decompressCfg.Enabled {
 			level.Debug(t.logger).Log("msg", "reading from compressed file", "filename", p)
-			decompressor, err := newDecompressor(t.metrics, t.logger, t.handler, t.positions, p, t.encoding)
+			decompressor, err := newDecompressor(t.metrics, t.logger, t.handler, t.positions, p, t.encoding, t.decompressCfg)
 			if err != nil {
 				level.Error(t.logger).Log("msg", "failed to start decompressor", "error", err, "filename", p)
 				continue
@@ -337,18 +342,6 @@ func (t *FileTarget) startTailing(ps []string) {
 		}
 		t.readers[p] = reader
 	}
-}
-
-func isCompressed(p string) bool {
-	ext := filepath.Ext(p)
-
-	for format := range supportedCompressedFormats() {
-		if ext == format {
-			return true
-		}
-	}
-
-	return false
 }
 
 // stopTailingAndRemovePosition will stop the tailer and remove the positions entry.